feat(be): split common lib into dedicated ones
