+++ backend/apps/client/src/auth/auth.controller.ts
@@ -10,11 +10,8 @@ import {
 } from '@nestjs/common'
 import { Request, Response } from 'express'
 import { REFRESH_TOKEN_COOKIE_OPTIONS, AUTH_TYPE } from '@libs/constants'
-import { AuthNotNeeded } from '@libs/decorator/auth-ignore.decorator'
-import {
-  InvalidJwtTokenException,
-  InvalidUserException
-} from '@libs/exception/business.exception'
+import { AuthNotNeeded } from '@libs/decorator'
+import { InvalidJwtTokenException, InvalidUserException } from '@libs/exception'
 import { AuthService } from './auth.service'
 import { LoginUserDto } from './dto/login-user.dto'
 import { AuthenticatedRequest } from './interface/authenticated-request.interface'
