+++ backend/apps/client/src/clarification/clarification.service.spec.ts
@@ -3,7 +3,7 @@ import { Test, type TestingModule } from '@nestjs/testing'
 import type { Clarification } from '@prisma/client'
 import { expect } from 'chai'
 import { stub } from 'sinon'
-import { EntityNotExistException } from '@libs/exception/business.exception'
+import { EntityNotExistException } from '@libs/exception'
 import { PrismaService } from '@libs/prisma'
 import { ContestService } from '@client/contest/contest.service'
 import { ClarificationService } from './clarification.service'
