+++ backend/apps/client/src/clarification/clarification.service.ts
@@ -1,7 +1,7 @@
 import { Injectable } from '@nestjs/common'
 import type { Clarification } from '@prisma/client'
 import { OPEN_SPACE_ID } from '@libs/constants'
-import { EntityNotExistException } from '@libs/exception/business.exception'
+import { EntityNotExistException } from '@libs/exception'
 import { PrismaService } from '@libs/prisma'
 import { ContestService } from '@client/contest/contest.service'
