+++ backend/apps/client/src/contest/contest-admin.controller.ts
@@ -16,13 +16,13 @@ import {
   UseGuards
 } from '@nestjs/common'
 import { type Contest, Role } from '@prisma/client'
-import { Roles } from '@libs/decorator/roles.decorator'
+import { Roles } from '@libs/decorator'
 import {
   ActionNotAllowedException,
   EntityNotExistException,
   UnprocessableDataException
-} from '@libs/exception/business.exception'
-import { CursorValidationPipe } from '@libs/pipe/cursor-validation.pipe'
+} from '@libs/exception'
+import { CursorValidationPipe } from '@libs/pipe'
 import { AuthenticatedRequest } from '@client/auth/interface/authenticated-request.interface'
 import { GroupLeaderGuard } from '@client/group/guard/group-leader.guard'
 import { RolesGuard } from '@client/user/guard/roles.guard'
