+++ backend/apps/client/src/notice/notice.controller.ts
@@ -9,9 +9,9 @@ import {
   InternalServerErrorException
 } from '@nestjs/common'
 import type { Notice } from '@prisma/client'
-import { AuthNotNeeded } from '@libs/decorator/auth-ignore.decorator'
-import { EntityNotExistException } from '@libs/exception/business.exception'
-import { CursorValidationPipe } from '@libs/pipe/cursor-validation.pipe'
+import { AuthNotNeeded } from '@libs/decorator'
+import { EntityNotExistException } from '@libs/exception'
+import { CursorValidationPipe } from '@libs/pipe'
 import { GroupMemberGuard } from '@client/group/guard/group-member.guard'
 import { RolesGuard } from '@client/user/guard/roles.guard'
 import type { UserNotice } from './interface/user-notice.interface'
