+++ backend/apps/client/src/notice/notice.service.spec.ts
@@ -3,7 +3,7 @@ import { Test, type TestingModule } from '@nestjs/testing'
 import type { Group, Notice } from '@prisma/client'
 import { expect } from 'chai'
 import { stub } from 'sinon'
-import { EntityNotExistException } from '@libs/exception/business.exception'
+import { EntityNotExistException } from '@libs/exception'
 import { PrismaService } from '@libs/prisma'
 import { GroupService } from '@client/group/group.service'
 import type { CreateNoticeDto } from './dto/create-notice.dto'
