+++ backend/apps/client/src/user/guard/roles.guard.ts
@@ -5,7 +5,7 @@ import {
 } from '@nestjs/common'
 import { Reflector } from '@nestjs/core'
 import { Role } from '@prisma/client'
-import { ROLES_KEY } from '@libs/decorator/roles.decorator'
+import { ROLES_KEY } from '@libs/decorator'
 import { UserService } from '@client/user/user.service'
 import type { AuthenticatedRequest } from '../../auth/interface/authenticated-request.interface'
