+++ backend/apps/client/src/workbook/workbook.controller.ts
@@ -9,9 +9,9 @@ import {
   Query
 } from '@nestjs/common'
 import type { Workbook } from '@prisma/client'
-import { AuthNotNeeded } from '@libs/decorator/auth-ignore.decorator'
-import { EntityNotExistException } from '@libs/exception/business.exception'
-import { CursorValidationPipe } from '@libs/pipe/cursor-validation.pipe'
+import { AuthNotNeeded } from '@libs/decorator'
+import { EntityNotExistException } from '@libs/exception'
+import { CursorValidationPipe } from '@libs/pipe'
 import { RolesGuard } from '@client/user/guard/roles.guard'
 import { GroupMemberGuard } from '../group/guard/group-member.guard'
 import { WorkbookService } from './workbook.service'
