+++ backend/libs/decorator/src/index.ts
@@ -1,0 +1,2 @@
+export * from './auth-ignore.decorator'
+export * from './roles.decorator'
