+++ backend/libs/exception/src/index.ts
@@ -1,0 +1,2 @@
+export * from './business.exception'
+export * from './controller.exception'
