+++ backend/libs/pipe/src/index.ts
@@ -1,0 +1,1 @@
+export * from './cursor-validation.pipe'
