+++ backend/nest-cli.json
@@ -53,13 +53,40 @@
         "tsConfigPath": "libs/cache/tsconfig.lib.json"
       }
     },
-    "common": {
+    "decorator": {
       "type": "library",
-      "root": "libs/common",
+      "root": "libs/decorator",
       "entryFile": "index",
-      "sourceRoot": "libs/common/src",
+      "sourceRoot": "libs/decorator/src",
       "compilerOptions": {
-        "tsConfigPath": "libs/common/tsconfig.lib.json"
+        "tsConfigPath": "libs/decorator/tsconfig.lib.json"
+      }
+    },
+    "exception": {
+      "type": "library",
+      "root": "libs/exception",
+      "entryFile": "index",
+      "sourceRoot": "libs/exception/src",
+      "compilerOptions": {
+        "tsConfigPath": "libs/exception/tsconfig.lib.json"
+      }
+    },
+    "pipe": {
+      "type": "library",
+      "root": "libs/pipe",
+      "entryFile": "index",
+      "sourceRoot": "libs/pipe/src",
+      "compilerOptions": {
+        "tsConfigPath": "libs/pipe/tsconfig.lib.json"
+      }
+    },
+    "constants": {
+      "type": "library",
+      "root": "libs/constants",
+      "entryFile": "index",
+      "sourceRoot": "libs/constants/src",
+      "compilerOptions": {
+        "tsConfigPath": "libs/constants/tsconfig.lib.json"
       }
     }
   }
