+++ backend/tsconfig.json
@@ -20,9 +20,12 @@
     "paths": {
       "@admin/*": ["apps/admin/src/*"],
       "@client/*": ["apps/client/src/*"],
-      "@libs/*": ["libs/common/src/*"],
       "@libs/prisma": ["libs/prisma/src/index.ts"],
-      "@libs/cache": ["libs/cache/src/index.ts"]
+      "@libs/cache": ["libs/cache/src/index.ts"],
+      "@libs/decorator": ["libs/decorator/src/index.ts"],
+      "@libs/exception": ["libs/exception/src/index.ts"],
+      "@libs/pipe": ["libs/pipe/src/index.ts"],
+      "@libs/constants": ["libs/constants/src/index.ts"]
     }
   },
   "ts-node": {
